% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blc.R
\name{blc}
\alias{blc}
\title{Fit a Boosting with LogitBoost Cost model}
\usage{
blc(train_x, train_y = NULL, n.tree = 500, n.minobsinnode = 10)
}
\arguments{
\item{train_x}{A data frame or a matrix of predictors}

\item{train_y}{A vector of response.}

\item{n.tree}{Number of trees to grow.}

\item{n.minobsinnode}{Minimum size of terminal nodes.}
}
\value{
A blc model
}
\description{
Fit a Boosting with LogitBoost Cost model
}
\examples{
library(EnsembleRankTrees)
# load the example data:
data(train_x_liver, train_y_liver)
model <- blc(train_x_liver, train_y_liver)

}
