% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rules.R
\name{extract.rules}
\alias{extract.rules}
\title{Extract rules containing comparing gene pairs from Random Rank Forest}
\usage{
extract.rules(train_x, train_y, n.trees = 500, mtry = 10, node_depth_max = 3)
}
\arguments{
\item{train_x}{A data frame or a matrix of predictors.}

\item{train_y}{A vector of response.}

\item{n.trees}{Number of trees to grow.}

\item{mtry}{Number of gene pairs randomly sampled as input at each split.}

\item{node_depth_max}{Maximum depth to which a tree should be grown. The default behavior is that this parameter is ignored.}
}
\value{
A list of rules extracted from the fitted Random Rank Forest model
}
\description{
Extract rules containing comparing gene pairs from Random Rank Forest
}
\examples{
library(EnsembleRankTrees)
# load the example data:
data(train_x_liver, train_y_liver)
# get training data and validation data
train_data <- split.train(train_x_liver, train_y_liver)$train.data
validation_data <- split.train(train_x_liver, train_y_liver)$validation.data
extract rules:
rules_table <- extract.rules(train_x_liver, train_y_liver)
}
