% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.blc.R
\name{predict.blc}
\alias{predict.blc}
\title{Prediction function for blc}
\usage{
\method{predict}{blc}(object, newdata)
}
\description{
Prediction function for blc
}
\examples{
predict(blc$model, liver_data$liver_test_x)
}
