% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rrf.R
\name{predict.rrf}
\alias{predict.rrf}
\title{Prediction function for rrf}
\usage{
\method{predict}{rrf}(object, newdata)
}
\description{
Prediction function for rrf
}
\examples{
predict(rrf$model, liver_data$liver_test_x)
}
