% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rules.R
\name{predict.rules}
\alias{predict.rules}
\title{Prediction function for rule-based methods}
\usage{
\method{predict}{rules}(data, decision.rules)
}
\arguments{
\item{data}{Test dataset}

\item{decision.rules}{The selected decision rules from select.rules}
}
\description{
Prediction function for rule-based methods
}
\examples{
predict.rules(liver_data$test_x_liver, selection)$Prediction.results
}
