% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf.R
\name{rrf}
\alias{rrf}
\title{Fit a Random Rank Forest model}
\usage{
rrf(train_x, train_y = NULL, n.tree = 500, mtry = NULL)
}
\arguments{
\item{train_x}{A data frame or a matrix of predictors.}

\item{train_y}{A vector of response. If omitted, rrf will run in unsupervised mode.}

\item{n.tree}{Number of trees to grow.}

\item{mtry}{Number of gene pairs randomly sampled as input at each split.}
}
\value{
A rrf model
}
\description{
Fit a Random Rank Forest model
}
\examples{
library(EnsembleRankTrees)
# load the example data:
data(train_x_liver, train_y_liver)
model <- rrf(train_x_liver, train_y_liver)
}
