% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rules.R
\name{select.rules}
\alias{select.rules}
\title{Select a set of rules that can be used to explain the model}
\usage{
select.rules(validation.data = validation_data, rule.table = rules_table)
}
\arguments{
\item{validation.data}{A separate validation data aimed to help select rules
from all decision rules obtained. Can be obtained from function split_data.}

\item{rule.table}{The object obtained from function extract_rule.}
}
\value{
A list of selected rules.

A plot of increase in accuracy on the validation dataset when adding rules to the ensemble.
}
\description{
Select a set of rules that can be used to explain the model
}
\examples{
selected_rules <- select.rule(validation_data, rules_table)$Selected.Rules
# show the line plot:
select.rule(validation_data, rules_table)$Plot.Num.Accuracy
}
