% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split.train}
\alias{split.train}
\title{Split a dataset into train and validation data, to be used in
extract_rules and select_rules}
\usage{
\method{split}{train}(train_x, train_y, seed_num = 1, valid_perc = 0.3)
}
\arguments{
\item{train_x}{A data frame or a matrix of predictors.}

\item{train_y}{A vector of response.}

\item{seed_num}{The seed number to be used for random sampling process.
Setting a fixed seed number ensures reproducibility of the results.}

\item{valid_perc}{The percentage of the original training data that
can be used a validation data. By default, 30% samples in the original
training data is used as validation data.}
}
\value{
A training dataset and a validation dataset (the last column are the
class labels)
}
\description{
Split a dataset into train and validation data, to be used in
extract_rules and select_rules
}
\examples{
library(EnsembleRankTrees)
# load the example data:
data(liver_data)
# get training data and validation data
train_data <- split.train(train_x_liver, train_y_liver)$train.data
validation_data <- split.train(train_x_liver, train_y_liver)$validation.data
}
